#16 Use int as tmdb and trakt ids instead of Long and String

+++ data/network/src/main/java/io/filmtime/data/network/trakt/AddHistoryRequest.kt
@@ -20,10 +20,10 @@ data class MovieHistory(
 
 @Serializable
 data class HistoryIDS(
-  val trakt: Long? = null,
-  val tvdb: Long? = null,
+  val trakt: Int? = null,
+  val tvdb: Int? = null,
   val imdb: String? = null,
-  val tmdb: Long? = null,
+  val tmdb: Int? = null,
   val slug: String? = null,
 )
